descriptors: move descriptor policy analysis into its own submodule
